requirements.txt
confusable-homoglyphs==2.0.2
dj-database-url==0.4.2
Django==1.11.7
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-imagekit==4.0.2
django-registration==2.3
django-tinymce==2.6.0
django-vote==2.1.5
gunicorn==19.7.1
olefile==0.44
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
